# src/main/java/com/gitlab/lbovolini/todo/user/api/PublicUserController.java
package com.gitlab.lbovolini.todo.user.api;

import com.gitlab.lbovolini.todo.user.service.UserService;
import com.gitlab.lbovolini.todo.common.model.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;

@RestController
@RequestMapping(
        consumes = MediaType.APPLICATION_JSON_VALUE,
        produces = MediaType.APPLICATION_JSON_VALUE)
public class PublicUserController {

    private final UserService userService;

    @Autowired
    public PublicUserController(UserService userService) {
        this.userService = userService;
    }

    @PostMapping("/register")
    public ResponseEntity<User> save(@Valid @RequestBody User user) {
        User savedUser = userService.register(user);

        return ResponseEntity.ok(savedUser);
    }

}